breakup spec file into individual files
